[Refactor] class명 변경, 이미지 추가 로직 수정
